Completeed both backend and frontend for add student ,list , view & edit .Bulk import half complted

+++ montfortuganda/src/main/java/com/erp/montfortuganda/student/entity/ErpStudentEnrollment.java
@@ -86,7 +86,7 @@ public class ErpStudentEnrollment {
 
     @NotNull(message = "Class is required")
     @Column(name = "class_id", nullable = false)
-    private Long classId;
+    private Integer classId;
 
     @Column(name = "section_id")
     private Long sectionId;
