+++ montfortuganda/src/main/java/com/erp/montfortuganda/student/entity/ErpStudentEnrollmentHistory.java
@@ -77,7 +77,7 @@ public class ErpStudentEnrollmentHistory {
 
     @NotNull(message = "Class is required")
     @Column(name = "class_id", nullable = false)
-    private Long classId;
+    private Integer classId;
 
     @Column(name = "section_id")
     private Long sectionId;
